+++ app/(dashboard)/dashboard/page.tsx
@@ -1,5 +1,7 @@
-import { Dashboard } from "@/components/dashboard"
+import { DashboardClient } from "@/components/dashboard-client"
 
 export default function DashboardPage() {
-  return <Dashboard />
+  // This is a server component that can fetch data if needed
+  // For now, we'll just render the client component
+  return <DashboardClient />
 }
