+++ app/(dashboard)/layout.tsx
@@ -1,10 +1,7 @@
-"use client"
-
-import { useEffect } from "react"
-import { useRouter, usePathname } from "next/navigation"
 import { AppSidebar } from "@/components/app-sidebar"
 import { AppHeader } from "@/components/app-header"
 import { SidebarInset } from "@/components/ui/sidebar"
+import { AuthGuard } from "@/components/auth-guard"
 
 const getPageTitle = (pathname: string): string => {
   const routes: Record<string, string> = {
@@ -22,28 +19,15 @@ export default function DashboardLayout({
 }: {
   children: React.ReactNode
 }) {
-  const router = useRouter()
-  const pathname = usePathname()
-
-  useEffect(() => {
-    // Check if user is authenticated
-    if (typeof window !== "undefined") {
-      const user = localStorage.getItem("user")
-      if (!user) {
-        router.push("/login")
-      }
-    }
-  }, [router])
-
   return (
-    <>
+    <AuthGuard>
       <AppSidebar />
       <SidebarInset>
-        <AppHeader title={getPageTitle(pathname)} />
+        <AppHeader />
         <main className="flex flex-1 flex-col gap-4 p-4 lg:gap-6 lg:p-6">
           {children}
         </main>
       </SidebarInset>
-    </>
+    </AuthGuard>
   )
 }
