+++ components/responsive-table.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import React from 'react'
 
 interface Column<T> {
