+++ components/sidebar-navigation-client.tsx
@@ -1,0 +1,9 @@
+"use client"
+
+import { usePathname } from "next/navigation"
+import { SidebarNavigation } from "@/components/sidebar-navigation"
+
+export function SidebarNavigationClient() {
+  const pathname = usePathname()
+  return <SidebarNavigation currentPath={pathname} />
+}
